refactor: Add Link in SideBar
