+++ components/mypage/SideBar/SideBar.tsx
@@ -1,14 +1,15 @@
+import Link from 'next/link';
 import Styled from './SideBar.styled';
 
 const SideBar = () => {
   const userId = '그루트 1';
   const sideBarList = {
     '신청한 약속': '/mypage/apply',
-    '신청받은 약속': 'mypage/receive',
-    '찜한 약속': `mypage/likes`,
-    '작성한 리뷰': `mypage/reviews`,
-    '작성한 포스트': `mypage/posts`,
-    '회원 정보 수정': `mypage/profile`,
+    '신청받은 약속': '/mypage/receive',
+    '찜한 포스트': `/mypage/likes`,
+    '작성한 포스트': `/mypage/posts`,
+    '회원 정보 수정': ``,
+    // '회원 정보 수정': `/mypage/profile`,
   };
   return (
     <Styled.aside>
@@ -18,8 +19,10 @@ const SideBar = () => {
             <Styled.h3>{`${userId}님!`}</Styled.h3>
             <Styled.h3>{`안녕하세요 :)`}</Styled.h3>
           </Styled.li>
-          {Object.keys(sideBarList).map(key => (
-            <Styled.li key={key}>{key}</Styled.li>
+          {Object.entries(sideBarList).map(([key, value]) => (
+            <Link key={key} href={value} passHref>
+              <Styled.li>{key}</Styled.li>
+            </Link>
           ))}
         </Styled.ul>
       </Styled.nav>
